Organize theme files (#679)
