+++ services/ui-src/src/components/pages/Dashboard/MobileDashboardTable.tsx
@@ -191,7 +191,7 @@ const AdminReleaseButton = ({
 
   return (
     <Button
-      variant="link"
+      variant="transparent"
       disabled={isDisabled}
       sx={sxOverride.adminActionButton}
       onClick={() => releaseReport!(report)}
@@ -214,7 +214,7 @@ const AdminArchiveButton = ({
         </Text>
       ) : (
         <Button
-          variant="link"
+          variant="transparent"
           sx={sxOverride.adminActionButton}
           onClick={() => archive(report)}
           disabled={report?.archived}
