+++ services/ui-src/src/styles/components/index.ts
@@ -1,0 +1,7 @@
+import buttonTheme from "./button";
+import linkTheme from "./link";
+
+export const components = {
+  Button: buttonTheme,
+  Link: linkTheme,
+};
