+++ services/ui-src/src/styles/foundations/breakpoints.ts
@@ -1,0 +1,8 @@
+export const breakpoints = {
+  // read this: https://bit.ly/3xSWnDt
+  base: "0em", // mobile (<=35em|560px)
+  sm: "35em", // tablet (>35em|560px and <=55em|880px)
+  md: "55em", // desktop, small (>55em|880px and <=75em|1200px)
+  lg: "75em", // desktop, large (>75em|1200px and <=100em|1600px)
+  xl: "100em", // desktop, ultrawide (>100em|1600px)
+};
