+++ services/ui-src/src/styles/foundations/fonts.ts
@@ -1,0 +1,4 @@
+export const fonts = {
+  heading: "Open Sans",
+  body: "Open Sans",
+};
