+++ services/ui-src/src/styles/foundations/sizes.ts
@@ -1,0 +1,16 @@
+export const sizes = {
+  appMax: "100vw",
+  basicPageWidth: "46rem",
+  reportPageWidth: "46rem",
+  // font sizes: https://design.cms.gov/utilities/font-size/
+  xs: "0.75rem", // 12px
+  sm: "0.875rem", // 14px
+  md: "1rem", // 16px
+  lg: "1.125rem", // 18px
+  xl: "1.3125rem", // 21px
+  "2xl": "1.5rem", // 24px
+  "3xl": "1.875rem", // 30px
+  "4xl": "2.25rem", // 36px
+  "5xl": "3rem", // 48px
+  "6xl": "3.75rem", // 60px
+};
